+++ R/markdown.R
@@ -156,20 +156,22 @@ eval_code_node <- function(node, env) {
   }
 
   chunk_opts <- utils::modifyList(
-    knitr_chunk_defaults,
+    knitr_chunk_defaults(),
     as.list(roxy_meta_get("knitr_chunk_options", NULL))
   )
 
   roxy_knit(text, env, chunk_opts)
 }
 
-knitr_chunk_defaults <- list(
-  error = FALSE,
-  fig.path = "man/figures/",
-  fig.process = function(path) basename(path),
-  comment = "#>",
-  collapse = TRUE
-)
+knitr_chunk_defaults <- function() {
+  list(
+    error = FALSE,
+    fig.path = "man/figures/",
+    fig.process = basename,
+    comment = "#>",
+    collapse = TRUE
+  )
+}
 
 str_set_all_pos <- function(text, pos, value, nodes) {
   # Cmark has a bug when reporting source positions for multi-line
@@ -424,12 +426,44 @@ mdxml_link_text <- function(xml_contents, state) {
   paste0(text, collapse = "")
 }
 
-mdxml_image = function(xml) {
+mdxml_image <- function(xml) {
   dest <- xml_attr(xml, "destination")
   title <- xml_attr(xml, "title")
+  fmt <- get_image_format(dest)
   paste0(
+    if (fmt == "html") "\\if{html}{",
+    if (fmt == "pdf") "\\if{pdf}{",
     "\\figure{", dest, "}",
-    if (nchar(title)) paste0("{", title, "}")
+    if (nchar(title)) paste0("{", title, "}"),
+    if (fmt %in% c("html", "pdf")) "}"
+  )
+}
+
+get_image_format <- function(path) {
+  should_restrict <- roxy_meta_get("restrict_image_formats") %||% TRUE
+  if (!should_restrict) {
+    return("all")
+  }
+
+  path <- tolower(path)
+  rx <- default_image_formats()
+  html <- grepl(rx$html, path)
+  pdf <- grepl(rx$pdf, path)
+  if (html && pdf) {
+    "all"
+  } else if (html) {
+    "html"
+  } else if (pdf) {
+    "pdf"
+  } else {
+    "all"
+  }
+}
+
+default_image_formats <- function() {
+  list(
+    html = "[.](jpg|jpeg|gif|png|svg)$",
+    pdf = "[.](jpg|jpeg|gif|png|pdf)$"
   )
 }