+++ NEWS.md
@@ -1,8 +1,8 @@
 # roxygen2 (development version)
 
-* `@describeIn` can be used to combine multiple types of functions
-  (generics, methods and other functions) into a `*.Rd` documentation file.
-  The resulting minidescription section organises the functions by type (#1181).
+* `@describeIn` can now be used to combine more types of functions
+  (generics, methods and other functions) into a single topic.
+  The resulting section organises the functions by type (#1181).
   Methods are recognized only if they extend the generic in the destination,
   or if the destination can heuristically be identified as a constructor.