+++ NEWS.md
@@ -1,5 +1,9 @@
 # roxygen2 (development version)
 
+* roxygen2 now only only includes PDF figures generated via markdown in
+  the PDF manual, and only includes SVG figures generated via markdown
+  in the HTML manual (#1399).
+
 * The new `knitr_chunk_options` option (in the `Roxygen` entry of
   `DESCRIPTION` or in `man/roxygen/meta.R`) is added to the knitr chunk
   options that roxygen2 uses for markdown code blocks and inline