+++ NEWS.md
@@ -1,3 +1,5 @@
+# roxygen2 (development version)
+
 # roxygen2 7.2.1
 
 ## Tags