+++ NEWS.md
@@ -1,5 +1,8 @@
 # roxygen2 (development version)
 
+* `@includeRmd` calls `local_reproducible_output()` to make code run in
+  included `.Rmd`s more consistent with other sources (#1431).
+
 # roxygen2 7.2.1
 
 ## Tags