+++ NEWS.md
@@ -3,6 +3,8 @@
 * `@includeRmd` calls `local_reproducible_output()` to make code run in
   included `.Rmd`s more consistent with other sources (#1431).
 
+* Fix duplicated argument in `roxy_block()` to avoid CRAN removal.
+
 # roxygen2 7.2.1
 
 ## Tags